refactor(board-plugins): clean up plugins
