+++ packages/components/src/area/boards/area.board.tsx
@@ -1,8 +1,9 @@
 import { createBoard } from '@wixc3/react-board';
-import { projectThemesPlugin } from '../../board-plugins/project-themes-plugin';
-import { Area } from '../area';
-import { AddFileIcon } from '../../icons';
 import React from 'react';
+import { projectThemesPlugin } from '../../board-plugins';
+import { AddFileIcon } from '../../icons';
+import { Area } from '../area';
+
 export default createBoard({
     name: 'Area',
     Board: () => (
