+++ packages/components/src/auto-complete/auto-complete.board.tsx
@@ -2,8 +2,7 @@ import { createBoard } from '@wixc3/react-board';
 import React from 'react';
 import type { ItemData } from '../board-assets';
 import { createItems, ItemRenderer } from '../board-assets';
-import { projectThemesPlugin } from '../board-plugins/project-themes-plugin';
-import { ZeeRootPlugin } from '../board-plugins/zee-root-plugin';
+import { projectThemesPlugin, ZeeRootPlugin } from '../board-plugins';
 import { AutoComplete } from './auto-complete';
 
 const items = createItems(30000);
