+++ packages/components/src/board-assets/index.ts
@@ -1,4 +1,5 @@
 export * from './items';
+export * from './utils';
 export * from './expandable-item-renderer';
 export * from './item-renderer';
 export * from './li-item-renderer';
