+++ packages/components/src/board-plugins/index.ts
@@ -1,0 +1,3 @@
+export * from './project-themes-plugin';
+export * from './zee-root-plugin';
+export * from './scenario-plugin/scenario-plugin';
