+++ packages/components/src/hooks/boards/use-scroll/use-scroll-with-ref.board.tsx
@@ -1,13 +1,7 @@
-import React from 'react';
 import { createBoard } from '@wixc3/react-board';
 import { expect } from 'chai';
-import {
-    expectElements,
-    expectElementText,
-    maxScroll,
-    scenarioPlugin,
-    scrollAction,
-} from '../../../board-plugins/scenario-plugin/scenario-plugin';
+import React from 'react';
+import { expectElements, expectElementText, maxScroll, scenarioPlugin, scrollAction } from '../../../board-plugins';
 import { ScrollHookSimulator } from './scroll-hook-simulator';
 
 export default createBoard({
