+++ packages/components/src/input/boards/input.board.tsx
@@ -1,6 +1,6 @@
-import React from 'react';
 import { createBoard } from '@wixc3/react-board';
-import { projectThemesPlugin } from '../../board-plugins/project-themes-plugin';
+import React from 'react';
+import { projectThemesPlugin } from '../../board-plugins';
 import { Input } from '../input';
 
 export default createBoard({
