+++ packages/components/src/scroll-list/boards/items-in-row.board.tsx
@@ -1,15 +1,15 @@
 import { createBoard } from '@wixc3/react-board';
 import React from 'react';
 import { createItems, getId, ItemRenderer } from '../../board-assets';
-import { projectThemesPlugin } from '../../board-plugins/project-themes-plugin';
 import {
     clickAction,
     expectElementsStyle,
     expectElementStyle,
     hoverAction,
+    projectThemesPlugin,
     scenarioPlugin,
     scrollAction,
-} from '../../board-plugins/scenario-plugin/scenario-plugin';
+} from '../../board-plugins';
 import { ScrollList } from '../scroll-list';
 
 const items = createItems();
